requirements.txt
discord.py
mutagen
pandas
PyNaCl
num2words